+++ src/application/gui/EventsVirtualListControl.hpp
@@ -22,7 +22,7 @@ class EventsVirtualListControl : public wxDataViewCtrl, public mvc::IView
      * @brief Sets the filtered events to be displayed in the control.
      * @param filteredIndices Vector of indices (or IDs) of events to display.
      */
-    void SetFilteredEvents(const std::vector<size_t>& filteredIndices);
+    void SetFilteredEvents(const std::vector<unsigned long>& filteredIndices);
 
     // implement IView interface
     virtual void OnDataUpdated() override;
